controller permission Gate call

+++ app/Http/Controllers/Backend/RoleController.php
@@ -18,7 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
-        // Gate::authorize('app.roles.index');
+        Gate::authorize('app.roles.index');
         $roles = Role::all();
         return view('backend.roles.index', compact('roles'));
     }
@@ -30,7 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
-        // Gate::authorize('app.roles.create');
+        Gate::authorize('app.roles.create');
         $modules = Module::all();
         return view('backend.roles.form', compact('modules'));
     }
@@ -43,7 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        // Gate::authorize('app.roles.create');
+        Gate::authorize('app.roles.create');
         $this->validate($request, [
             'name'  =>  'required|unique:roles',
             'permissions'   =>  'required|array',
@@ -76,7 +76,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        // Gate::authorize('app.roles.edit');
+        Gate::authorize('app.roles.edit');
         $modules = Module::all();
         return view('backend.roles.form', compact('role', 'modules'));
     }
@@ -90,7 +90,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
-        // Gate::authorize('app.roles.edit');
+        Gate::authorize('app.roles.edit');
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name),
@@ -108,7 +108,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
-        // Gate::authorize('app.roles.destroy');
+        Gate::authorize('app.roles.destroy');
         if ($role->deletable) {
             $role->delete();
             notify()->success("Role Successfully Deleted", "Deleted");
